requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.12
django-cleanup==6.0.0
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2==2.7.7
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
